platform/mowgli: Limit slot1 to Gen3 speed
When link reset/retrain, the previous method cannot maintain the max-link-speed limit,
delete it and use the method provided by Frederic:
Add the "ibm, maximum link speed" attribute to the PHB device tree at index 0.
The phb4 code already looks for it and it will set up the link correctly.

Signed-off-by: LuluTHSu <Lulu_Su@wistron.com>